botlib: close F34's call site with a gated call
The .so's BotChooseBestFightWeapon calls F34; the DLL's calls the thin
Linux build was calling the wrong evaluator.  Now gated -- the only place in
this tree where a build gate sits INSIDE a function body rather than around a
whole function -- and our resolved call list for that function is byte-for-byte
the real one's: None, AAS_Time, F34, Q_stricmp, EA_UseItem, AAS_Time.

THE AUDIT IS BLIND TO THIS, AND THAT IS THE POINT WORTH REMEMBERING.
since it encodes layout -- so retargeting a call changes no compared byte.  No
row moved.  My earlier note predicting this would "close the ELF row" was wrong
for exactly that reason: BotChooseBestFightWeapon's OUR+2 is unrelated, being
one extra 4-byte frame slot, two zero-inits in the opposite order, and a `for`
CALL LISTS, never against insn_diffs.

Also fixes a header slip from the previous commit: F34's prototype was placed
after be_ai_weight.h's closing include guard, so it was re-declared on every
inclusion.  Swept all 34 botlib headers for the same mistake -- none others.

  msvc6:  MATCH 756, ZERO rows moved             (the #else path is unchanged)

Source-commit: 75eab84017d9d5b221abd9b23b45a752380a6c6f

+++ botlib/be_ai_weight.c
@@ -836,7 +836,7 @@ void __cdecl InterbreedWeightConfigs(weightconfig_t *a, weightconfig_t *b)
  * The blend divides two ints and only then converts, exactly as FuzzyWeight_r
  * does (`idiv` then `fild`, not `fdiv`) -- a step function, not a smooth
  * interpolation.  Faithful; see .claude/memory/fuzzy_weight_system.md. */
-float __cdecl F34(int *inventory, weight_t *w)
+double __cdecl F34(int *inventory, weight_t *w)
 {
   fuzzyseperator_t *fs;
   float w1, w2;
